refactor: Moved features panel
